+++ client.go
@@ -88,6 +88,11 @@ func (r *Blob) Unmarshal(b []byte) error { r.Content = b; return nil }
 // Marshal implements part of the Object interface
 func (r *Blob) Marshal() ([]byte, error) { return r.Content, nil }
 
+// netconn is the interface for a riak connection
+type netconn interface {
+	io.ReadWriteCloser
+}
+
 // conn is a connection
 type conn struct {
 	*net.TCPConn
@@ -119,11 +124,6 @@ func (t *timerconn) Write(b []byte) (int, error) {
 	n, err := t.conn.Write(b)
 	t.twait += time.Since(now)
 	return n, err
-}
-
-// netconn is the interface for a riak connection
-type netconn interface {
-	io.ReadWriteCloser
 }
 
 // write wraps the TCP write
@@ -169,7 +169,6 @@ func Dial(addrs []string, clientID string) (*Client, error) {
 	cl := &Client{
 		tag:   0,
 		id:    []byte(clientID),
-		pool:  new(sync.Pool),
 		addrs: naddrs,
 	}
 
@@ -204,7 +203,7 @@ func (c *Client) Close() {
 	// we'll hang if we don't make
 	// the connection pool immediately
 	// GC-able
-	c.pool = nil
+	c.pool = sync.Pool{}
 
 	runtime.GC()
 	nspin := 0
@@ -222,20 +221,6 @@ func (c *Client) Close() {
 
 func (c *Client) closed() bool {
 	return atomic.LoadInt32(&c.tag) == 1
-}
-
-// Client represents a pool of connections
-// to a Riak cluster.
-type Client struct {
-	conns int32 // total live conns
-	pad1  [4]byte
-	inuse int32 // conns in use
-	pad2  [4]byte
-	tag   int32 // 0 = open; 1 = closed
-	pad3  [4]byte
-	id    []byte
-	pool  *sync.Pool
-	addrs []*net.TCPAddr
 }
 
 // can we add another connection?
@@ -393,33 +378,6 @@ func readResponse(c *conn, b []byte) ([]byte, byte, error) {
 		b = append(b, scratch[:nn]...)
 	}
 	return b[5:n], b[4], err
-}
-
-// send the contents of a buffer and receive a response
-// back in the same buffer
-func (c *Client) doBuf(code byte, msg []byte) ([]byte, byte, error) {
-	node, err := c.popConn()
-	if err != nil {
-		return nil, 0, err
-	}
-
-	msg[4] = code
-	_, err = node.Write(msg)
-	if err != nil {
-		c.err(node)
-		return nil, 0, err
-	}
-	if err != nil {
-		c.err(node)
-		return nil, 0, err
-	}
-	msg, code, err = readResponse(node, msg)
-	if err == nil {
-		c.done(node)
-	} else {
-		c.err(node)
-	}
-	return msg, code, nil
 }
 
 func (c *Client) req(msg protom, code byte, res unmarshaler) (byte, error) {
